refactor(db): reorganize db utils
