--- packages/backend/src/lib/index.ts
@@ -1,3 +1,0 @@
-export {dynamodb} from './dynamodb-client';
-export {getPhotos} from './photosTable';
-export type {DynamoPhotoEntity, PhotoEntity} from './photosTable';
